Add validation for missing content-type header.
We require the POST data to be in JSON format and that the content-type header be set to application/json.
This PR does that validation and returns a decent message. Previously the error message was a None type error which is not helpful to anyone.

Thank you to @bambrp for bring this to our attention. I've closed your PR, https://github.com/alphagov/notifications-api/pull/2652, in favour of this one as I have included some unit tests and updated the post_template endpoint because it had to be handled specially since the data can be empty for that request.